Add mark read/unread actions to chapter dialog
This commit enhances the chapter long-press dialog by adding "Mark as read" and "Mark as unread" buttons.

To support this, the `libraryId` has been added to the `VolumeDetailPayload` and is now passed from the series detail screen to the volume detail screen. The dialog UI has been updated to accommodate the new actions.